+++ docs/manual/purity-checker.tex
@@ -10,7 +10,7 @@ All checkers utilize purity annotations on called methods.
 You do not need to run the Purity Checker directly.
 However you may run just the Purity Checker by
 supplying the following command-line options to javac:
-\code{-processor org.checkerframework.util.PurityChecker}.
+\code{-processor org.checkerframework.common.purity.PurityChecker}.
 
 
 \sectionAndLabel{Purity annotations}{purity-annotations}
