+++ framework/src/main/java/org/checkerframework/common/purity/PurityAnnotatedTypeFactory.java
@@ -1,4 +1,4 @@
-package org.checkerframework.framework.util;
+package org.checkerframework.common.purity;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
@@ -7,6 +7,7 @@ import java.util.Set;
 import org.checkerframework.common.basetype.BaseAnnotatedTypeFactory;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 
+/** The Purity Checker is used to determine if methods are deterministic and side-effect-free. */
 public class PurityAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
     public PurityAnnotatedTypeFactory(BaseTypeChecker checker) {
