+++ framework/src/main/java/org/checkerframework/common/purity/PurityChecker.java
@@ -1,0 +1,24 @@
+package org.checkerframework.common.purity;
+
+import java.util.LinkedHashSet;
+import org.checkerframework.common.basetype.BaseTypeChecker;
+import org.checkerframework.common.basetype.BaseTypeVisitor;
+
+/**
+ * Perform purity checking only.
+ *
+ * @checker_framework.manual #type-refinement-purity Side effects, determinism, purity, and
+ *     flow-sensitive analysis
+ */
+public class PurityChecker extends BaseTypeChecker {
+
+    @Override
+    protected BaseTypeVisitor<?> createSourceVisitor() {
+        return new PurityVisitor(this);
+    }
+
+    @Override
+    protected LinkedHashSet<Class<? extends BaseTypeChecker>> getImmediateSubcheckerClasses() {
+        return new LinkedHashSet<>();
+    }
+}
