+++ framework/src/main/java/org/checkerframework/common/purity/PurityUnqualified.java
@@ -1,4 +1,4 @@
-package org.checkerframework.framework.util;
+package org.checkerframework.common.purity;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
